+++ simple_history/admin.py
@@ -6,6 +6,7 @@ from django.contrib.admin import helpers
 from django.contrib.admin.utils import unquote
 from django.contrib.auth import get_permission_codename, get_user_model
 from django.core.exceptions import PermissionDenied
+from django.core.paginator import Paginator
 from django.shortcuts import get_object_or_404, render
 from django.urls import re_path, reverse
 from django.utils.encoding import force_str
@@ -21,6 +22,7 @@ SIMPLE_HISTORY_EDIT = getattr(settings, "SIMPLE_HISTORY_EDIT", False)
 class SimpleHistoryAdmin(admin.ModelAdmin):
     object_history_template = "simple_history/object_history.html"
     object_history_form_template = "simple_history/object_history_form.html"
+    history__list_per_page = 50
 
     def get_urls(self):
         """Returns the additional urls used by the Reversion admin."""
@@ -60,6 +62,9 @@ class SimpleHistoryAdmin(admin.ModelAdmin):
             except action_list.model.DoesNotExist:
                 raise http.Http404
 
+        paginator = Paginator(action_list, self.history__list_per_page)
+        action_list_page = paginator.get_page(request.GET.get("page"))
+
         if not self.has_view_history_or_change_history_permission(request, obj):
             raise PermissionDenied
 
@@ -67,7 +72,7 @@ class SimpleHistoryAdmin(admin.ModelAdmin):
         for history_list_entry in history_list_display:
             value_for_entry = getattr(self, history_list_entry, None)
             if value_for_entry and callable(value_for_entry):
-                for list_entry in action_list:
+                for list_entry in action_list_page.object_list:
                     setattr(list_entry, history_list_entry, value_for_entry(list_entry))
 
         content_type = self.content_type_model_cls.objects.get_for_model(
@@ -80,7 +85,7 @@ class SimpleHistoryAdmin(admin.ModelAdmin):
         )
         context = {
             "title": self.history_view_title(request, obj),
-            "action_list": action_list,
+            "action_list": action_list_page,
             "module_name": capfirst(force_str(opts.verbose_name_plural)),
             "object": obj,
             "root_path": getattr(self.admin_site, "root_path", None),