+++ CHANGES.rst
@@ -41,6 +41,7 @@ Unreleased
 - Dropped support for Django 4.0, which reached end-of-life on 2023-04-01 (gh-1202)
 - Added support for Django 4.2 (gh-1202)
 - Made ``bulk_update_with_history()`` return the number of model rows updated (gh-1206)
+- Added pagination to ``SimpleHistoryAdmin`` (gh-1220)
 - Fixed ``HistoryRequestMiddleware`` not cleaning up after itself (i.e. deleting
   ``HistoricalRecords.context.request``) under some circumstances (gh-1188)
 - Made ``HistoryRequestMiddleware`` async-capable (gh-1209)